feat: redesign page structure and user creation flow
- Move users table to dedicated /users route
- Repurpose index as main menu
- Change user creation from automatic to manual trigger
